refactor(container): move app_container to infrastructure
